+++ spec/rubocop/cop/style/method_call_parentheses_spec.rb
@@ -52,6 +52,16 @@ describe RuboCop::Cop::Style::MethodCallParentheses do
     end
   end
 
+  it 'registers an offense for `obj.method ||= func()`' do
+    inspect_source(cop, 'obj.method ||= func()')
+    expect(cop.offenses.size).to eq 1
+  end
+
+  it 'registers an offense for `obj.method &&= func()`' do
+    inspect_source(cop, 'obj.method &&= func()')
+    expect(cop.offenses.size).to eq 1
+  end
+
   it 'auto-corrects by removing unneeded braces' do
     new_source = autocorrect_source(cop, 'test()')
     expect(new_source).to eq('test')